chore: refactor all sections into their own component files
