+++ frontend/routes/auth/settings/settings-section.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export const SettingsSection = ({ children }: { children: ReactNode }) => (
+  <div className="pt-6 border-t border-1 border-vega-dark-150 pb-6">{children}</div>
+)
